1st game
This is a 2 player game where a player rolls the dice and a random die between 1 to 6 is displayed, if the number 1 is chosen the active player forfeits their points and the next player begins their turn. if the hold button Is chosen then they keep all their points and the next player plays. if the new game button is chosen then the game resets and player one begins the game.