GuessTheNumber
This is a guessing game where the player has to guess which random number between 1-20 has been chosen by the program. If the wrong number has been chosen then a clue would be given to the player if the number is too low or too high, the player has 20 chances to pick the right number. A high score is also there to keep track of the players highest score.